+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `accountOptionsButtonBehavior` prop to the `login` interface. It defaults to `popover` and can be changed to `link`
+- New `accountOptionsButtonBehavior` prop to the site-editor
+
 ## [2.25.0] - 2020-03-10
 
 ### Fixed